feat(cli): add installed AgentOps evidence commands
Expose orlando-agentops-quickstart, orlando-agentops-eval, and
orlando-agentops-gcp-preflight so a clean wheel install can prove the
fake HTTP path without a checkout.
