+++ src/agentops/cli/__init__.py
@@ -1,0 +1,1 @@
+"""Installed AgentOps evidence CLIs (synthetic fixtures; no provider key)."""
